+++ changelog.md
@@ -4,6 +4,7 @@ Saves from 2.3 are not compatible with 2.4.
 
 ## Features/Improvements
 
+* **[Flight Planner]** Air-to-air and SEAD escorts will no longer be automatically planned for packages that are not in range of threats.
 * **[Plugins]** [The EWRS plugin](https://github.com/Bob7heBuilder/EWRS) is now included.
 
 ## Fixes